requirements.txt
colorama==0.4.6 
contourpy==1.2.0 
cycler==0.12.1 
fonttools==4.44.3 
kiwisolver==1.4.5 
llvmlite==0.41.1 
matplotlib==3.8.1 
numba==0.58.1 
numpy==1.26.2 
packaging==23.2 
pandas==2.1.3 
pillow==10.1.0 
pyparsing==3.1.1 
python-dateutil==2.8.2 
pytz==2023.3.post1 
six==1.16.0 
tqdm==4.66.1 
tzdata==2023.3 
